Update background processor

+++ BE/src/main/java/com/gcp/practise/parking/services/ParkingLotService.java
@@ -2,10 +2,12 @@ package com.gcp.practise.parking.services;
 
 import com.gcp.practise.parking.dtos.responses.ParkingSpotResponse;
 import com.gcp.practise.parking.dtos.responses.ReservationResponse;
+
+
 import java.util.List;
 
 public interface ParkingLotService {
     List<ReservationResponse> getCurrentReservations();
-    ParkingSpotResponse findParkingSpotById(Long spotId);
+    ParkingSpotResponse findParkingSpotById(Integer spotId);
     List<ParkingSpotResponse> getAllSpots();
 }
